+++ src/modules/admin/widgets/grids/columns/AssetThumbnailColumn.php
@@ -4,20 +4,12 @@ namespace davidhirtz\yii2\cms\hotspot\modules\admin\widgets\grids\columns;
 
 use davidhirtz\yii2\cms\models\Asset;
 use davidhirtz\yii2\skeleton\helpers\Html;
-use Yii;
 
 class AssetThumbnailColumn extends \davidhirtz\yii2\cms\modules\admin\widgets\grids\columns\AssetThumbnailColumn
 {
-    protected function renderDataCellContent($model, $key, $index): string
+    protected function renderThumbnail(Asset $model, int $key, int $index): string
     {
-        if (!$model->file->hasPreview()) {
-            return '';
-        }
-
-        $content = Html::tag('div', '', [
-            'style' => 'background-image:url(' . ($model->file->getTransformationUrl('admin') ?: $model->file->getUrl()) . ');',
-            'class' => 'thumb',
-        ]);
+        $content = parent::renderThumbnail($model, $key, $index);
 
         if ($hotspotCount = $model->getAttribute('hotspot_count')) {
             $badge = Html::tag('div', $hotspotCount, [
@@ -31,10 +23,6 @@ class AssetThumbnailColumn extends \davidhirtz\yii2\cms\modules\admin\widgets\gr
             ]);
         }
 
-        $isLink = $model instanceof Asset && Yii::$app->getUser()->can($model->isEntryAsset() ? 'entryAssetUpdate' : 'sectionAssetUpdate', [
-            'asset' => $model,
-        ]);
-
-        return $isLink ? Html::a($content, ['/admin/cms/asset/update', 'id' => $model->id]) : $content;
+        return $content;
     }
 }