+++ tests/config/test.php
@@ -1,33 +1,36 @@
 <?php
 
-use davidhirtz\yii2\cms\hotspot\Bootstrap;
-use yii\web\Session;
+use davidhirtz\yii2\media\Bootstrap;
+
+if (is_file(__DIR__ . '/db.php')) {
+    require(__DIR__ . '/db.php');
+}
 
 return [
-    'id' => 'yii2-cms',
-    'aliases' => [
-        // This is a fix for the broken aliasing of `BaseMigrateController::getNamespacePath()`
-        '@davidhirtz/yii2/cms/hotspot' => __DIR__ . '/../../src/',
-    ],
     'bootstrap' => [
         Bootstrap::class,
     ],
     'components' => [
-        'assetManager' => [
-            'linkAssets' => true,
-        ],
         'db' => [
             'dsn' => getenv('MYSQL_DSN') ?: 'mysql:host=127.0.0.1;dbname=yii2_cms_hotspot_test',
             'username' => getenv('MYSQL_USER') ?: 'root',
             'password' => getenv('MYSQL_PASSWORD') ?: '',
             'charset' => 'utf8',
-            ...is_file(__DIR__ . '/db.php') ? require (__DIR__ . '/db.php') : [],
         ],
-        'mailer' => [
-            'useFileTransport' => true,
-        ],
-        'session' => [
-            'class' => Session::class,
+    ],
+    'modules' => [
+        'media' => [
+            'transformations' => [
+                'xs' => [
+                    'width' => 100,
+                ],
+                'sm' => [
+                    'width' => 200,
+                ],
+                'md' => [
+                    'width' => 300,
+                ],
+            ],
         ],
     ],
     'params' => [